chore: use custom BottomSheet to safe snapPoints
